feat(HTD-72): setup integration overview page
- refactor IntegrationGrid, make the component isolate without props
