--- src/components/sections/integration/grid/grid.types.ts
@@ -1,5 +1,0 @@
-import type { IntegrationCardProps } from '@/components/modules/integration-card/integration-card.types';
-
-export interface IntegrationGridProps {
-  integrations: IntegrationCardProps[];
-}
